+++ src/GSvar/SvWidget.h
@@ -3,7 +3,10 @@
 
 #include <QWidget>
 #include <QTableWidgetItem>
+#include <QByteArray>
+#include <QByteArrayList>
 #include "SvList.h"
+#include "BedPeFile.h"
 
 namespace Ui {
 	class SvWidget;
@@ -15,20 +18,36 @@ class SvWidget : public QWidget
 	Q_OBJECT
 
 public:
-	explicit SvWidget(const QString& file_name, QWidget *parent = 0);
+	explicit SvWidget(const QStringList& bedpe_file_paths, QWidget *parent = 0);
 
 private:
 	Ui::SvWidget *ui;
 
 	SvList svs_;
 
+	BedpeFile sv_bedpe_file_;
+
+	///List of annotations which are shown in the widget
+	QByteArrayList annotations_to_show_;
+
 	void addInfoLine(const QString text);
 
+	///load bedpe data file and set display
 	void loadSVs(const QString& file_name);
 
-	void filterAnnotationsForNumber(QByteArray anno_name, double filter_thresh, QBitArray &pass);
-
 	void disableGui(const QString& message);
+
+	///Returns column index of main QTableWidget svs_ by Name, -1 if not in widget
+	int colIndexbyName(const QString& name);
+
+	///File widgets with data from INFO_A and INFO_B column
+	void setInfoWidgets(const QByteArray& name, int row, QTableWidget* widget);
+
+	void clearGUI();
+
+	///resize widgets to cell content
+	void resizeQTableWidget(QTableWidget* table_widget);
+
 
 signals:
 	void openSvInIGV(QString coords);
@@ -37,10 +56,19 @@ private slots:
 	///copy filtered SV table to clipboard
 	void copyToClipboard();
 
+	///load new SV file if other bedpe file is selected
+	void fileNameChanged();
+
 	///update SV table if filter for types was changed
 	void filtersChanged();
 
 	void SvDoubleClicked(QTableWidgetItem* item);
+
+	///update SV details and INFO widgets
+	void SvSelectionChanged();
+
+	///Context menu that shall appear if right click on variant
+	void showContextMenu(QPoint pos);
 };
 
 #endif // SVWIDGET_H