+++ src/SeqPurge/OutputWorker.cpp
@@ -4,8 +4,8 @@ OutputWorker::OutputWorker(QList<AnalysisJob>& job_pool, QString out1, QString o
 	: QRunnable()
 	, terminate_(false)
 	, job_pool_(job_pool)
-	, ostream1(new FastqOutfileStream(out1))
-	, ostream2(new FastqOutfileStream(out2))
+	, ostream1(new FastqOutfileStream(out1, params.compression_level))
+	, ostream2(new FastqOutfileStream(out2, params.compression_level))
 	, ostream3()
 	, ostream4()
 	, params_(params)
@@ -13,8 +13,8 @@ OutputWorker::OutputWorker(QList<AnalysisJob>& job_pool, QString out1, QString o
 {
 	if (out3_base.trimmed()!="")
 	{
-		ostream3.reset(new FastqOutfileStream(out3_base + "_R1.fastq.gz"));
-		ostream4.reset(new FastqOutfileStream(out3_base + "_R2.fastq.gz"));
+		ostream3.reset(new FastqOutfileStream(out3_base + "_R1.fastq.gz", params.compression_level));
+		ostream4.reset(new FastqOutfileStream(out3_base + "_R2.fastq.gz", params.compression_level));
 	}
 }